selftests/bpf: Fix testing narrow ctx load on 32-bit
Subtests for invalid narrow loads in test verifier_ctx use u32 loads which
are only relevant for 64-bit systems. Update macro narrow_load() to build
code with u16 loads, which apply also on 32-bit systems.

Fixes: ba578b87fe2b ("selftests/bpf: Test invalid narrower ctx load")

+++ tools/testing/selftests/bpf/progs/verifier_ctx.c
@@ -225,7 +225,7 @@ __naked void null_check_8_null_bind(void)
 	__naked void invalid_narrow_load##ctx##field(void)		\
 	{								\
 		asm volatile ("						\
-		r1 = *(u32 *)(r1 + %[off]);				\
+		r1 = *(u16 *)(r1 + %[off]);				\
 		r0 = 0;							\
 		exit;"							\
 		:							\
